Part 2

+++ src/main/java/com/example/demo/controller/ProductController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.beans.IntrospectionException;
@@ -34,25 +35,33 @@ public class ProductController {
 
     @PatchMapping("/{id}")
     public ResponseEntity<Product> updateProduct(@PathVariable("id") int id, @RequestBody Product productDetails) throws IntrospectionException {
-        Optional<Product> productData = productRepository.findById(id);
+        if (isAdmin()){
+            Optional<Product> productData = productRepository.findById(id);
 
-        if (productData.isPresent()) {
-            Product product = productData.get();
-            BeanUtils.copyProperties(productDetails, product, getIgnoredProperties(productDetails));
-            product.setUpdatedAt(Instant.now());
-            return new ResponseEntity<>(productRepository.save(product), HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            if (productData.isPresent()) {
+                Product product = productData.get();
+                BeanUtils.copyProperties(productDetails, product, getIgnoredProperties(productDetails));
+                product.setUpdatedAt(Instant.now());
+                return new ResponseEntity<>(productRepository.save(product), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        }else {
+            return  new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> deleteProduct(@PathVariable("id") int id) {
-        try {
-            productRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        if (isAdmin()) {
+            try {
+                productRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        }else {
+            return  new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
@@ -64,19 +73,24 @@ public class ProductController {
 
     @PostMapping
     public ResponseEntity<Product> createProduct(@RequestBody Product product) {
-        try {
-            // Set createdAt and updatedAt timestamps
-            Instant now = Instant.now();
-            product.setCreatedAt(now);
-            product.setUpdatedAt(now);
+        if (isAdmin()) {
+            try {
+                // Set createdAt and updatedAt timestamps
+                Instant now = Instant.now();
+                product.setCreatedAt(now);
+                product.setUpdatedAt(now);
 
-            // Save the product to the database
-            Product savedProduct = productRepository.save(product);
+                // Save the product to the database
+                Product savedProduct = productRepository.save(product);
 
-            return new ResponseEntity<>(savedProduct, HttpStatus.CREATED);
-        } catch (Exception e) {
-            // Handle exceptions (e.g., database errors)
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+                return new ResponseEntity<>(savedProduct, HttpStatus.CREATED);
+            } catch (Exception e) {
+                // Handle exceptions (e.g., database errors)
+                return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        }
+        else {
+            return  new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
@@ -92,6 +106,17 @@ public class ProductController {
         try {
             return Objects.isNull(new PropertyDescriptor(propertyName, source.getClass()).getReadMethod().invoke(source));
         } catch (Exception e) {
+            return false;
+        }
+    }
+
+    private boolean isAdmin(){
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (email.equals("admin@admin.com")){
+            System.out.println("email" +email);
+            return true;
+        }
+        else{
             return false;
         }
     }
